% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{avg}
\alias{avg}
\title{Batting Average}
\usage{
avg(df)
}
\arguments{
\item{df}{dataframe containing "h" & "ab"}
}
\value{
new column containing batting average.
}
\description{
Batting Average
}
\examples{
avg(playerSample)

}
