% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{bb_pct}
\alias{bb_pct}
\title{Walk Rate (BB\%)}
\usage{
bb_pct(df)
}
\arguments{
\item{df}{dataframe containing season totals "pa" & "bb"}
}
\value{
new column containing "bb_pct".
}
\description{
Walk Rate (BB\%)
}
\examples{
bb_pct(playerSample)

}
