% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{k_pct}
\alias{k_pct}
\title{Strikeout Rate (K\%)}
\usage{
k_pct(df)
}
\arguments{
\item{df}{dataframe containing season totals "k" & "bb"}
}
\value{
new column containing "k_pct".
}
\description{
Strikeout Rate (K\%)
}
\examples{
k_pct(playerSample)

}
