% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_stats.R
\name{long_form}
\alias{long_form}
\title{Convert rolling stats to long format for shiny plotting}
\usage{
long_form(df)
}
\arguments{
\item{df}{dataframe built by function \code{rolling_stats()}containing game by game rolling season stats.}
}
\value{
dataframe containing data transformed into long format
}
\description{
Convert rolling stats to long format for shiny plotting
}
