% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{makeTable}
\alias{makeTable}
\title{Table workflow}
\usage{
makeTable(df)
}
\arguments{
\item{df}{dataframe containing game by game stats ideally identical in form to \link{playerSample}.}
}
\value{
dataframe containing a season summary.
}
\description{
Table workflow
}
\examples{
makeTable(playerSample)

}
