% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{obp}
\alias{obp}
\title{On base percentage}
\usage{
obp(df)
}
\arguments{
\item{df}{dataframe containing season totals for "h", "bb", "hbp","ab" & "sac"}
}
\value{
new column containing batting average.
}
\description{
On base percentage
}
\examples{
obp(playerSample)

}
