% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{ops}
\alias{ops}
\title{OPS}
\usage{
ops(df)
}
\arguments{
\item{df}{dataframe containing season totals "obp" & "slg"}
}
\value{
new column containing "ops".
}
\description{
OPS
}
\examples{
ops(playerSample)

}
