% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{playerSample}
\alias{playerSample}
\title{Sample Player data}
\format{
dataframe with one row per game
\itemize{
\item \code{date} - Date data type including year to ensure accurate calculations. Date of each game.
\item \code{pa} - Plate appearances
\item \code{ab} - At bats
\item \code{r} - Runs
\item \code{h} - Hits
\item \code{rbi} - Runs batted in
\item \code{x2b} - Doubles
\item \code{x3b} - Triples
\item \code{hr} - Home runs
\item \code{sac} - Sacrifice bunts/flies
\item \code{bb} - Walks
\item \code{k} - Strikeouts
\item \code{hbp} - Hit by pitch
\item \code{roe} - Reached on error
\item \code{fc} - Fielder's choice
\item \code{sb} - Stolen bases
\item \code{cs} - Caught stealing
}
}
\source{
Sample dataset for simulation
}
\usage{
playerSample
}
\description{
Sample data to demonstrate the correct columns and data types for player data.
}
\examples{
data("playerSample", package = "hitplotr")
head(playerSample)

}
\keyword{datasets}
