% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_single}
\alias{plot_single}
\title{Plotting function for dashboard}
\usage{
plot_single(df)
}
\arguments{
\item{df}{dataframe containing long form game by game data from the \code{long_form} function.}
}
\value{
time series plot with rolling values colored by stat.
}
\description{
Plotting function for dashboard
}
