% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHits.R
\name{plotHits}
\alias{plotHits}
\title{Main Plotting Function}
\usage{
plotHits(df)
}
\arguments{
\item{df}{dataframe containing game by game player data in the same format as \link{playerSample}}
}
\value{
a shiny app with summarized statistics organized in a table and an interactive cumulative plot.
}
\description{
Main Plotting Function
}
\examples{
\dontrun{
plotHits(playerSample)}

}
