% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{rolling_plot}
\alias{rolling_plot}
\title{Pulling together functions to put the plotting workflow in one function}
\usage{
rolling_plot(df)
}
\arguments{
\item{df}{dataframe containing wide form game by game data, similar structure to \link{playerSample}.}
}
\value{
time series plot with rolling values colored by stat.
}
\description{
Pulling together functions to put the plotting workflow in one function
}
