% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_stats.R
\name{rolling_stats}
\alias{rolling_stats}
\title{Calculate rolling stats}
\usage{
rolling_stats(df)
}
\arguments{
\item{df}{dataframe containing game by game stats, cleaned by the \code{standardize_names()} function.}
}
\value{
new dataframe containing rolling stats for each game since the beginning of data.
}
\description{
Calculate rolling stats
}
\examples{
rolling_stats(playerSample)

}
