% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{sb_pct}
\alias{sb_pct}
\title{Stolen Base \%}
\usage{
sb_pct(df)
}
\arguments{
\item{df}{dataframe containing season totals "sb" & "cs"}
}
\value{
new column containing "sbp".
}
\description{
Stolen Base \%
}
\examples{
sb_pct(playerSample)

}
