% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{season_summary}
\alias{season_summary}
\title{Season Totals}
\usage{
season_summary(df)
}
\arguments{
\item{df}{dataframe containing game by game stats.}
}
\value{
new dataframe containing a season summary.
}
\description{
Season Totals
}
\examples{
season_summary(playerSample)

}
