% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{season_totals}
\alias{season_totals}
\title{Calculate Season Totals}
\usage{
season_totals(df)
}
\arguments{
\item{df}{dataframe of game by game statistics.}
}
\value{
Dataframe with aggregated statistics
}
\description{
Calculate Season Totals
}
