% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{slg}
\alias{slg}
\title{Slugging}
\usage{
slg(df)
}
\arguments{
\item{df}{dataframe containing season totals "h", "x2b", "x3b","hr" & "ab"}
}
\value{
new column containing "slg".
}
\description{
Slugging
}
\examples{
slg(playerSample)

}
