% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{standardize_names}
\alias{standardize_names}
\title{Adjust Common Baseball Terminology to Integrate into Functions}
\usage{
standardize_names(df)
}
\arguments{
\item{df}{data.frame containing all the required columns \link{playerSample}}
}
\value{
data.frame with standardized column names
}
\description{
Adjust Common Baseball Terminology to Integrate into Functions
}
